feat: for user_session tests, move dict fixtures to conftest and clean up test_db
